refactoring sudoku names for pylint
